win32

+++ barretenberg/cpp/src/barretenberg/client_ivc/client_ivc.cpp
@@ -26,7 +26,7 @@ ClientIVC::ClientIVC(size_t num_circuits, TraceSettings trace_settings)
     // Allocate BN254 commitment key based on the max dyadic Mega structured trace size and translator circuit size.
     // https://github.com/AztecProtocol/barretenberg/issues/1319): Account for Translator only when it's necessary
     size_t commitment_key_size =
-        std::max(trace_settings.dyadic_size(), 1UL << TranslatorFlavor::CONST_TRANSLATOR_LOG_N);
+        std::max(trace_settings.dyadic_size(), static_cast<size_t>(1) << TranslatorFlavor::CONST_TRANSLATOR_LOG_N);
     info("BN254 commitment key size: ", commitment_key_size);
     bn254_commitment_key = CommitmentKey<curve::BN254>(commitment_key_size);
 }
